--- bot-server/src/config/server.ts
@@ -1,3 +1,0 @@
-import express from "express";
-
-export const server = express();
