--- bot-server/src/data/data.js
@@ -1,113 +1,0 @@
-export const studentContents = [
-  {
-    tim: {
-      nama: "Esa",
-      no: 4,
-    },
-    siswa: [
-      "Brilliant Ozora Himmatana",
-      "Rafa Chadhijah Syafitri",
-      "Ahmad Danial Adzimmi",
-      "Vanya Namira Aisyah",
-      "Rangga Pratama",
-      "Rahma Cahya Kamila",
-      "Dewi Retna Arum Sasi",
-    ],
-  },
-  {
-    tim: {
-      nama: "Brilliant Ozora David Muklis",
-      no: 5,
-    },
-    siswa: [
-      "Teguh Sugiharto Santoso",
-      "Nikmatus Solihah",
-      "Rasya Eka Ayu Yulistya",
-      "Marsha Hera Ayunda",
-      "Fahmi Nur Yudisyah",
-      "Dafan Zain Pratama",
-    ],
-  },
-  {
-    tim: {
-      nama: "Vanya",
-      no: 1,
-    },
-    siswa: [
-      "Firdaus Andhika Pamungkas",
-      "Yora Nisbunan",
-      "Salva Vidominna",
-      "Aura Septiarsya",
-      "Dilla Ayuhan",
-      "Dewi Afifatuz Zahro",
-      "Zahran Zaidan Nasution",
-    ],
-  },
-  {
-    tim: {
-      nama: "Bintang",
-      no: 2,
-    },
-    siswa: [
-      "Andaraka Putra Risriyana",
-      "Najwa Istighfara",
-      "Aryya Agata Dikawasistha",
-      "Erlyta Zeva Ditya",
-      "Rodifatuz Zamaniah",
-      "Daffa Aziz Ghiffari",
-      "Daifan Nur Amali",
-    ],
-  },
-  {
-    tim: {
-      nama: "Dewi A",
-      no: 3,
-    },
-    siswa: [
-      "Irma Lisnawati",
-      "Esa Raditya Mantovany",
-      "Bintang Rafif Avecinna",
-      "Muhammad Alliffian Fakhri Effendi",
-      "Zahra Marjannah",
-      "Navila El Khairina",
-      "Putri Velani",
-    ],
-  },
-];
-
-export const students = [
-  "Firdaus Andhika Pamungkas",
-  "Yora Nisbunan",
-  "Salva Vidominna",
-  "Aura Septiarsya",
-  "Dilla Ayuhan",
-  "Dewi Afifatuz Zahro",
-  "Zahran Zaidan Nasution",
-  "Andaraka Putra Risriyana",
-  "Najwa Istighfara",
-  "Aryya Agata Dikawasistha",
-  "Erlyta Zeva Ditya",
-  "Rodifatuz Zamaniah",
-  "Daffa Aziz Ghiffari",
-  "Daifan Nur Amali",
-  "Irma Lisnawati",
-  "Esa Raditya Mantovany",
-  "Bintang Rafif Avecinna",
-  "Muhammad Alliffian Fakhri Effendi",
-  "Zahra Marjannah",
-  "Navila El Khairina",
-  "Putri Velani",
-  "Brilliant Ozora Himmatana",
-  "Rafa Chadhijah Syafitri",
-  "Ahmad Danial Adzimmi",
-  "Vanya Namira Aisyah",
-  "Rangga Pratama",
-  "Rahma Cahya Kamila",
-  "Dewi Retna Arum Sasi",
-  "Teguh Sugiharto Santoso",
-  "Nikmatus Solihah",
-  "Rasya Eka Ayu Yulistya",
-  "Marsha Hera Ayunda",
-  "Fahmi Nur Yudisyah",
-  "Dafan Zain Pratama",
-];
