--- bot-server/src/models/admin/adminBroadcaster.type.ts
@@ -1,9 +1,0 @@
-export interface AdminBroadcaster {
-  id: string;
-  admin_id: string;
-  pfpslide_enabled: boolean;
-  pfpslide: string;
-  pfpslide_interval: number;
-  created_at: Date;
-  updated_at: Date;
-}
