--- bot-server/src/models/admin/adminSettings.type.ts
@@ -1,8 +1,0 @@
-export interface AdminSettings {
-  id: string;
-  admin_id: string;
-  bot_delay_enabled: boolean;
-  bot_delay: null;
-  created_at: Date;
-  updated_at: Date;
-}
