--- bot-server/src/models/api-responses/adminDetail.type.ts
@@ -1,8 +1,0 @@
-import { AdminGroup } from "../../admin/admin-group";
-
-export interface AdminDetailResponse {
-  success: boolean;
-  msg: string;
-  group: AdminGroup;
-  status: number;
-}
