--- bot-server/src/models/api-responses/adminShuffle.type.ts
@@ -1,6 +1,0 @@
-export interface AdminShufflePfpResponse {
-  success: boolean;
-  msg: string;
-  status: number;
-  image: string;
-}
