--- bot-server/src/models/api-responses/allGroup.type.ts
@@ -1,8 +1,0 @@
-import { Group } from "../../group/group";
-
-export interface AllGroupsResponse {
-  success: boolean;
-  msg: string;
-  groups: Group[];
-  status: number;
-}
