--- bot-server/src/models/api-responses/apiCheck.type.type.ts
@@ -1,5 +1,0 @@
-export interface APICheckResponse {
-  success: boolean;
-  msg: string;
-  status: number;
-}
