--- bot-server/src/models/api-responses/checkGroupPw.type.ts
@@ -1,6 +1,0 @@
-export interface CheckIfGroupHasPwResponse {
-  success: boolean;
-  msg: string;
-  value: boolean;
-  status: number;
-}
