--- bot-server/src/models/api-responses/checkGroupRegistered.type.ts
@@ -1,6 +1,0 @@
-export interface CheckIfGroupRegisteredResponse {
-  success: boolean;
-  msg: string;
-  value: boolean;
-  status: number;
-}
