--- bot-server/src/models/api-responses/groupDetail.type.ts
@@ -1,8 +1,0 @@
-import { Group } from "../../group/group";
-
-export interface GroupDetailResponse {
-  success: boolean;
-  msg: string;
-  group: Group;
-  status: number;
-}
