--- bot-server/src/models/api-responses/groupShuffle.type.ts
@@ -1,6 +1,0 @@
-export interface GroupShufflePfpResponse {
-  success: boolean;
-  msg: string;
-  status: number;
-  image: string;
-}
