--- bot-server/src/models/api-responses/otpLogin.type.ts
@@ -1,6 +1,0 @@
-export interface OTPLoginResponse {
-  success: boolean;
-  msg: string;
-  otp: number;
-  status: number;
-}
