--- bot-server/src/models/group/groupSettings.type.ts
@@ -1,8 +1,0 @@
-export interface GroupSettings {
-  id: string;
-  group_id: string;
-  lock_mention_everyone: boolean;
-  schedule_piket: boolean;
-  created_at: Date;
-  updated_at: Date;
-}
