--- bot-server/src/models/group/kelompok.type.ts
@@ -1,4 +1,0 @@
-export interface GroupKelompok {
-  participants: string[];
-  numberOfParticipants: number;
-}
