--- bot-server/src/models/participant/category.type.ts
@@ -1,4 +1,0 @@
-export interface Category {
-  id: string;
-  localized_display_name: string;
-}
