--- bot-server/src/models/participant/id.type.ts
@@ -1,9 +1,0 @@
-export interface ID {
-  user: string;
-  server: Server;
-  _serialized: string;
-}
-
-export enum Server {
-  CUs = "c.us",
-}
