--- bot-server/src/models/participant/profileOptions.type.ts
@@ -1,4 +1,0 @@
-export interface ProfileOptions {
-  cartEnabled: boolean;
-  commerceExperience: string;
-}
