--- bot-server/src/models/participant/typeEnum.type.ts
@@ -1,3 +1,0 @@
-export enum Type {
-  In = "in",
-}
