--- bot-server/src/models/participant/website.type.ts
@@ -1,3 +1,0 @@
-export interface Website {
-  url: string;
-}
