--- bot-server/src/utils/utils.ts
@@ -1,8 +1,0 @@
-// Seed Generator
-const generateRandomSeed = () => {
-  let randomNumber = "";
-  for (let i = 0; i < 16; i++) {
-    randomNumber += Math.floor(Math.random() * 10);
-  }
-  return randomNumber;
-};
